Revert "Fix engagement end during de-authentication"
This reverts commit 5d25c7704725430c5556584fbd3aa9c642431f4a.
